+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix severity_in_level SQL function [#1312](https://github.com/greenbone/gvmd/pull/1312)
 
 ### Removed
+- Remove solution element from VT tags [#886](https://github.com/greenbone/gvmd/pull/886)
 - Drop GMP scanners [#1269](https://github.com/greenbone/gvmd/pull/1269)
 - Reduce Severity Classes [#1285](https://github.com/greenbone/gvmd/pull/1285)
 - Removed Severity Classes [#1288](https://github.com/greenbone/gvmd/pull/1288)